+++ functions.php
@@ -360,6 +360,7 @@ require_once 'admin/_main.php';         // Admin Settings.
 require_once 'assets/symbolicons.php';  // Symbolicons/Font Icons.
 require_once 'blocks/_main.php';        // Custom Blocks.
 require_once 'debugger/_main.php';      // Debugger and data fixers.
+require_once 'of-the-day/_main.php';    // Of the Day RSS code
 require_once 'plugins/_main.php';       // Tweaks for Plugins.
 require_once 'rest-api/_main.php';      // Our Rest API.
 require_once 'statistics/_main.php';    // Stats.
